perf: move host check modal outside & improve host protocol verification
